call_notes_app/requirements.txt
sounddevice==0.5.1
numpy==2.2.3
amazon-transcribe==0.6.2
boto3>=1.35.0
python-docx==1.1.2
